Update for more setup

+++ osx/setup.sh
@@ -2,7 +2,18 @@
 
 script_path=${0:a:h}
 
-rm ~/.zshrc ~/.p10k.zsh
+if [[ -a ~/.zshrc || -a ~/.p10k.zsh ]]; then
+    rm ~/.zshrc ~/.p10k.zsh
+fi
+
+if [[ ! -a ~/.oh-my-zsh ]]; then
+    sh -c "$(curl -fsSL https://raw.github.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+    git clone --depth=1 https://github.com/romkatv/powerlevel10k.git $ZSH_CUSTOM/themes/powerlevel10k
+fi
+
 
 ln -s $script_path/.zshrc ~/.zshrc
 ln -s $script_path/.p10k.zsh ~/.p10k.zsh
+
+git config --global user.name andrew.greene
+git config --global user.email andrew@noreply.com
